@@ -15,22 +15,23 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "^3.1.0",
-    "@docusaurus/plugin-content-docs": "^3.1.0",
-    "@docusaurus/plugin-ideal-image": "^3.1.0",
-    "@docusaurus/preset-classic": "^3.1.0",
-    "@docusaurus/remark-plugin-npm2yarn": "^3.1.0",
-    "@docusaurus/theme-common": "^3.1.0",
-    "@docusaurus/utils": "^3.1.0",
+    "@docusaurus/core": "^3.1.1",
+    "@docusaurus/plugin-content-docs": "^3.1.1",
+    "@docusaurus/plugin-ideal-image": "^3.1.1",
+    "@docusaurus/preset-classic": "^3.1.1",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.1.1",
+    "@docusaurus/theme-common": "^3.1.1",
+    "@docusaurus/utils": "^3.1.1",
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-brands-svg-icons": "^6.5.1",
     "@fortawesome/free-regular-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mdx-js/react": "^3.0.0",
+    "@iconify/react": "^4.1.1",
+    "@mdx-js/react": "^3.0.1",
     "@popperjs/core": "^2.11.8",
-    "@swc/core": "^1.3.102",
-    "@types/react-dom": "^18.2.18",
+    "@swc/core": "^1.4.1",
+    "@types/react-dom": "^18.2.19",
     "clsx": "^2.1.0",
     "docusaurus-plugin-plausible": "^0.0.5",
     "docusaurus-plugin-sass": "^0.2.5",
@@ -44,19 +45,33 @@
     "remark-definition-list": "^2.0.0",
     "remark-deflist": "^1.0.0",
     "remark-math": "6",
-    "sass": "^1.69.7",
-    "swc-loader": "^0.2.3"
+    "sass": "^1.70.0",
+    "swc-loader": "^0.2.6"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^3.1.0",
-    "@docusaurus/tsconfig": "^3.1.0",
-    "@types/jest": "^29.5.11",
-    "@types/react": "^18.2.47",
+    "@docusaurus/module-type-aliases": "^3.1.1",
+    "@docusaurus/tsconfig": "^3.1.1",
+    "@docusaurus/types": "^3.1.1",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.17",
+    "@types/react": "^18.2.55",
+    "@typescript-eslint/eslint-plugin": "7.0.1",
+    "@typescript-eslint/parser": "7.0.1",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-formatjs": "4.12.2",
+    "eslint-plugin-jsx-a11y": "6.8.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-react": "7.33.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-regexp": "2.2.0",
     "jest": "^29.7.0",
+    "prettier": "3.2.5",
+    "prettier-plugin-css-order": "2.0.1",
     "prop-types": "^15.8.1",
     "react-waypoint": "^10.3.0",
     "typescript": "^5.3.3",
-    "webpack": "^5.89.0"
+    "webpack": "^5.90.1"
   },
   "browserslist": {
     "production": [
@@ -73,5 +88,10 @@
   "engines": {
     "node": ">=18.14"
   },
-  "packageManager": "yarn@4.0.2"
+  "packageManager": "yarn@4.1.0",
+  "dependenciesMeta": {
+    "@docusaurus/tsconfig@3.1.1": {
+      "unplugged": true
+    }
+  }
 }
