@@ -1,6 +1,10 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
-
+// Import Types
+import type { Config } from '@docusaurus/types';
+import type * as Preset from '@docusaurus/preset-classic';
+import type {Options as DocsOptions} from '@docusaurus/plugin-content-docs';
+import type {Options as PageOptions} from '@docusaurus/plugin-content-pages';
 // Import the Docusaurus version.
 const { DOCUSAURUS_VERSION } = require("@docusaurus/utils");
 // Setup our Prism themes.
@@ -58,7 +62,7 @@ const commonDocsPluginConfig = {
 import plausiblePlugin from 'docusaurus-plugin-plausible';
 
 /** @type {import('@docusaurus/types').Config} */
-const config = {
+const config: Config = {
   title: 'Docusaurus.community',
   tagline: 'Docusaurus community knowledge sharing and plugin directory.',
   favicon: 'img/favicon.ico',
@@ -76,31 +80,31 @@ const config = {
     [
       'classic',
       /** @type {import('@docusaurus/preset-classic').Options} */
-      ({
+      {
         docs: {
           routeBasePath: 'knowledge',
-          sidebarPath: require.resolve('./sidebars.js'),
+          sidebarPath: './sidebars.js',
           // Please change this to your repo.
           // Remove this to remove the "edit this page" links.
           editUrl:
             'https://github.com/homotechsual/docusaurus.community/edit/main/',
           ...commonDocsPluginConfig,
-        },
+        } satisfies DocsOptions,
         blog: false,
         pages: {
           ...admonitionsConfig,
           ...remarkPluginsConfig
-        },
+        } satisfies PageOptions,
         theme: {
-          customCss: require.resolve('./src/scss/custom.scss'),
+          customCss: './src/scss/custom.scss',
         },
-      }),
+      } satisfies Preset.Options,
     ],
   ],
 
   themeConfig:
     /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
-    ({
+    {
       // Replace with your project's social card
       image: 'img/docusaurus-social-card.jpg',
       colorMode: {
@@ -114,6 +118,7 @@ const config = {
           alt: 'Docusaurus.community Logo',
           src: 'img/logo.svg',
         },
+        hideOnScroll: true,
         items: [
           {
             type: 'doc',
@@ -218,7 +223,7 @@ const config = {
         darkTheme: darkCodeTheme,
         additionalLanguages: ['powershell','bash','docker', 'diff', 'json']
       },
-    }),
+    } satisfies Preset.ThemeConfig,
   plugins: [
     [
       '@docusaurus/plugin-ideal-image',
@@ -244,15 +249,15 @@ const config = {
         id: 'contributing',
         path: 'contributing',
         routeBasePath: 'contributing',
-        sidebarPath: require.resolve('./sidebars.js'),
+        sidebarPath: './sidebars.js',
         editUrl: 'https://github.com/homotechsual/docusaurus.community/tree/main/',
         ...commonDocsPluginConfig,
-      },
+      } satisfies DocsOptions,
     ]
   ],
   webpack: {
     jsLoader: (isServer: boolean) => ({
-      loader: require.resolve('swc-loader'),
+      loader: 'swc-loader',
       options: {
         jsc: {
           parser: {
@@ -280,7 +285,6 @@ const config = {
       headingIds: false,
       admonitions: false,
     }
-  }
+  },
 };
-
-module.exports = config;
+export default config;
